refactor: introduce abstraction layer for slice-by-slice measurements
