+++ src/clearcontrol/microscope/lightsheet/postprocessing/containers/SliceBySliceDCTS2DContainer.java
@@ -13,38 +13,12 @@ import org.apache.commons.math.stat.descriptive.moment.Mean;
  * 05 2018
  */
 
-public class SliceBySliceDCTS2DContainer extends DataContainerBase {
+public class SliceBySliceDCTS2DContainer extends SliceBySliceMeasurementContainer {
 
-    double[] mMeasurements = null;
-    Double mMeanMeasurement = null;
-
-    public SliceBySliceDCTS2DContainer(long pTimePoint, double pX, double pY, double pZ, double[] pDCTS2D) {
-        super(pTimePoint);
-        mMeasurements = new double[pDCTS2D.length];
-        System.arraycopy(pDCTS2D, 0, mMeasurements, 0, mMeasurements.length);
+    public SliceBySliceDCTS2DContainer(long pTimePoint, double[] pDCTS2D) {
+        super(pTimePoint, pDCTS2D);
     }
 
-    @Override
-    public boolean isDataComplete() {
-        return true;
-    }
-
-    @Override
-    public void dispose() {
-    }
-
-    public double[] getMeasurements() {
-        double[] lMeasurements = new double[mMeasurements.length];
-        System.arraycopy(mMeasurements, 0, lMeasurements, 0, mMeasurements.length);
-        return lMeasurements;
-    }
-
-    public double getMeanMeasurement() {
-        if (mMeanMeasurement == null) {
-            mMeanMeasurement = new Mean().evaluate(mMeasurements);
-        }
-        return mMeanMeasurement;
-    }
 }
